+++ R/get_upheno.R
@@ -5,7 +5,7 @@
 #' Unified Phenotype Ontology (uPheno)}.
 #'
 #' @param file Can be one of the following:
-#' \itemize{
+#' \describe{
 #' \item{"ontology"}{Creates an \link[simona]{ontology_DAG} R object by
 #'  importing the OBO file directly from the official
 #'  \href{https://github.com/obophenotype/upheno}{uPheno GitHub repository}.}