+++ R/map_upheno_data_i2.R
@@ -10,7 +10,9 @@ map_upheno_data_i2 <- function(pheno_map_method,
     hgnc_label2 <- hgnc_id2 <- gene <- gene_label1 <- gene_label2 <-
     n_genes_db1 <- object <- gene_label <- db <- .  <-
     n_genes_db2 <- subject_taxon_label1 <- subject_taxon_label2 <-
-    phenotype_genotype_score <- equivalence_score <- NULL;
+    phenotype_genotype_score <- equivalence_score <- subject_category <- 
+    object_category <- category <- taxon_label <- gene_taxon_label1 <- 
+    gene_taxon_label2 <- NULL;
   
   pheno_map_method <- pheno_map_method[1]
   gene_map_method <- gene_map_method[1]